Fix: move token tracking to meta events (#1379)
* Fix: move token tracking to meta events

* Fix 0 tokens
